#227 allow spatial dependence
